dispatcher/requirements.txt
Flask==2.3.3
requests==2.31.0
google-auth==2.27.0
gunicorn==21.2.0